CU-8695edtv5 feature
Moved landing page board game list to a seperate controller action
